Refactor update pack parsing; new pack v4
- Changed the initialization of the current working directory to point directly to the "assets" subdirectory.
- Removed unnecessary restoration of the original current working directory after initialization.
- Updated the `pack` function signature to return additional block metadata.
- Modified the reading of the DLC pack to use a file handle instead of reading raw bytes directly.
- Removed unused `read_pack_bytes` method to streamline code.
- Cleaned up the DLC pack test by removing the unused `pack_key` parameter.

+++ src/macros.rs
@@ -20,7 +20,7 @@
 #[doc(hidden)]
 macro_rules! human_bytes {
     ($bytes:expr) => {{
-        let bytes: usize = $bytes;
+        let bytes: u64 = $bytes as u64;
         const UNITS: [&str; 5] = ["B", "KB", "MB", "GB", "TB"];
         let mut size = bytes as f64;
         let mut unit = 0;
@@ -29,7 +29,7 @@ macro_rules! human_bytes {
             unit += 1;
         }
         if unit == 0 {
-            format!("{} {}", size as usize, UNITS[unit])
+            format!("{} {}", size as u64, UNITS[unit])
         } else {
             format!("{:.2} {}", size, UNITS[unit])
         }
